requirement.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
aiohttp==3.9.1
requests==2.31.0